+++ documentation/changelog.md
@@ -34,6 +34,7 @@ description: List of all the changes and current versions.
 + Added Various Interface Descriptions
 + Remove Dead Code for Help
 + Fixed Non-Unique Identifiers for Stream Based Data
++ Fix ownership bug where no data is present on vehicle
 ```
 
 ## 2.0.3